Fix lzfw_write return code.
Should be ssize_t, since the return is bytes written and the
interface is 64-bit.

+++ zfswrap/libzfswrap.h
@@ -370,7 +370,7 @@ int lzfw_read(lzfw_vfs_t *p_vfs, creden_t *p_cred, lzfw_vnode_t *p_vnode, void *
  * @param offset: the offset to write
  * @return 0 in case of success, the error code overwise
  */
-int lzfw_write(lzfw_vfs_t *p_vfs, creden_t *p_cred, lzfw_vnode_t *p_vnode, void *p_buffer, size_t size, int behind, off_t offset);
+ssize_t lzfw_write(lzfw_vfs_t *p_vfs, creden_t *p_cred, lzfw_vnode_t *p_vnode, void *p_buffer, size_t size, int behind, off_t offset);
 
 /**
  * Vectorwise write to file
